+++ R/converge_ok.R
@@ -113,3 +113,9 @@ is_singular.glmmTMB <- function(x, tolerance = 1e-5, ...) {
   vc <- collapse_cond(lme4::VarCorr(x))
   any(sapply(vc, function(.x) any(abs(diag(.x)) < tolerance)))
 }
+
+#' @export
+is_singular.default <- function(x, ...) {
+  FALSE
+}
+